ML_week_02/defect_detection/requirements.txt
# ── Core ML & CV ──────────────────────────────────
ultralytics>=8.1.0
torch>=2.1.0
torchvision>=0.16.0
onnx>=1.15.0
onnxruntime>=1.17.0

# ── Data & Image Processing ──────────────────────
numpy>=1.26.0
pandas>=2.1.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
albumentations>=1.3.1
scikit-learn>=1.4.0

# ── API ──────────────────────────────────────────
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.5.0
aiofiles>=23.2.0

# ── UI ───────────────────────────────────────────
streamlit>=1.30.0
gradio>=4.14.0
plotly>=5.18.0

# ── Monitoring & Logging ─────────────────────────
tensorboard>=2.15.0
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.2

# ── Utilities ────────────────────────────────────
requests>=2.31.0
gdown>=5.1.0
httpx>=0.26.0
Jinja2>=3.1.3
openpyxl>=3.1.2

# ── Testing ──────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
